refactoring of parsers
