+++ src/main/java/xyz/rtsvk/alfax/webserver/endpoints/DirectMessageEndpoint.java
@@ -5,6 +5,7 @@ import discord4j.core.GatewayDiscordClient;
 import discord4j.core.object.entity.User;
 import discord4j.core.object.entity.channel.PrivateChannel;
 import xyz.rtsvk.alfax.util.Database;
+import xyz.rtsvk.alfax.webserver.IEndpoint;
 import xyz.rtsvk.alfax.webserver.Request;
 
 import java.util.List;
