+++ src/main/java/xyz/rtsvk/alfax/webserver/endpoints/EditMessageEndpoint.java
@@ -4,6 +4,7 @@ import discord4j.common.util.Snowflake;
 import discord4j.core.GatewayDiscordClient;
 import discord4j.core.object.entity.Message;
 import xyz.rtsvk.alfax.util.Database;
+import xyz.rtsvk.alfax.webserver.IEndpoint;
 import xyz.rtsvk.alfax.webserver.Request;
 
 import java.util.List;
